Fix for type mismatches that showed up on 64-bit platforms

+++ dynchk/libs/libc/recvmsg.c
@@ -4,9 +4,9 @@
 #include <dlfcn.h>
 #include <sys/socket.h>
 #undef recvmsg
-static int(*funcptr) (int , struct msghdr * , int ) = 0;
+static ssize_t(*funcptr) (int , struct msghdr * , int ) = 0;
 
-int recvmsg (int arg0 , struct msghdr * arg1 , int arg2 )
+ssize_t recvmsg (int arg0 , struct msghdr * arg1 , int arg2 )
 {
 	if(!funcptr)
 		funcptr = dlsym(RTLD_NEXT, "recvmsg");
@@ -16,7 +16,7 @@ int recvmsg (int arg0 , struct msghdr * arg1 , int arg2 )
 	return funcptr(arg0, arg1, arg2);
 }
 
-int lsb_recvmsg (int arg0 , struct msghdr * arg1 , int arg2 )
+ssize_t lsb_recvmsg (int arg0 , struct msghdr * arg1 , int arg2 )
 {
 	if(!funcptr)
 		funcptr = dlsym(RTLD_NEXT, "recvmsg");
